moved guest update functions to api.py and added more tests
