+++ tests/test_validators.py
@@ -9,6 +9,10 @@ from vagrantfile_builder.validators import (
 )
 
 
+def guest_data():
+    return copy.deepcopy(mock_guest_data['guests'][0])
+
+
 def mock_guest_remove_key(popme):
     guest = copy.deepcopy(mock_guest_data['guests'][0])
     return guest.pop(popme)
@@ -24,19 +28,21 @@ def test_validate_missing_required_key_name_raises_attribute_error(value):
 
 
 def test_validate_missing_required_key_vagrant_box_name_raises_attribute_error():
-    guest = copy.deepcopy(mock_guest_data['guests'][0])
+    guest = guest_data()
     guest['vagrant_box'].pop('name')
     with pytest.raises(AttributeError):
-        validate_required_keys(mock_guest_remove_key('name'))
+        validate_required_keys(guest)
 
 
 def test_validate_empty_required_name_raises_value_error():
-    mock_guest_data['guests'][0]['name'] = ''
+    guest = guest_data()
+    guest['name'] = ''
     with pytest.raises(ValueError):
-        validate_required_values(mock_guest_data['guests'][0])
+        validate_required_values(guest)
 
 
 def test_validate_empty_required_vagrant_box_name_raises_value_error():
-    mock_guest_data['guests'][0]['vagrant_box']['name'] = ''
+    guest = guest_data()
+    guest['vagrant_box']['name'] = ''
     with pytest.raises(ValueError):
-        validate_required_values(mock_guest_data['guests'][0])
+        validate_required_values(guest)
