+++ vagrantfile_builder/api.py
@@ -1,15 +1,20 @@
+import copy
+
 from .custom_filters import (
     explode_port,
 )
 
 from .loaders import (
     render_from_template,
+    load_data,
 )
 
 from .constants import (
     TEMPLATE_DIR,
     BLACKHOLE_LOOPBACK_MAP,
+    ALL_GUEST_DEFAULTS,
 )
+
 
 custom_filters = [
     explode_port,
@@ -35,7 +40,59 @@ def generate_loopbacks(host_list=None, network='127.255.1'):
     return {**host_to_loopback_map, **BLACKHOLE_LOOPBACK_MAP}
 
 
-def update_interfaces(total_interfaces, interface_list):
+def update_context(source, target):
+    """
+    Take a source dict and update it with target data
+    :param source: Source data dictionary
+    :param target: Target data dictionary to update
+    :return: Return new dict with merged data
+    """
+    new_context = copy.deepcopy(target)
+    for k, v in source.items():
+        if k in target:
+            if isinstance(target[k], dict):
+                new_context[k].update(source[k])
+            else:
+                new_context[k] = source[k]
+
+    return new_context
+
+
+def update_guest_data(
+        guest_data,
+        guest_defaults='./guest-defaults.yml',
+        all_guest_defaults=ALL_GUEST_DEFAULTS):
+    """
+    Build data vars for guests. This function will take all_guest_defaults and merge in
+    guest and guest group vars.
+    :param guest_data: Dict of guest data
+    :param guest_defaults: Location of guest defaults file
+    :param all_guest_defaults: All guest default data
+    :return: Updated Dict of guest data
+    """
+    try:
+        guest_defaults = load_data(guest_defaults)
+    except FileNotFoundError:
+        print(f'File: {guest_defaults} not found')
+        guest_defaults = {}
+
+    new_guest_data = {'guests': []}
+    for guest in guest_data['guests']:
+        # Copy the default dict, that contains all top level variables
+        # Group vars, then host vars will be merged into this dict
+        default_context = copy.deepcopy(all_guest_defaults['guest_defaults'])
+        if guest_defaults and guest['vagrant_box'].get('name') in guest_defaults:
+            # Merge group vars with host vars
+            group_context = guest_defaults.get(guest['vagrant_box']['name'])
+            new_context = update_context(group_context, default_context)
+            new_guest_data['guests'].append(update_context(guest, new_context))
+        else:
+            # No group vars found, just merge host vars
+            new_guest_data['guests'].append(update_context(guest, default_context))
+    return new_guest_data
+
+
+def add_blackhole_interfaces(total_interfaces, interface_list):
     """
     Adds blackhole interfaces to host data by inserting a
     dict of blackhole config in the correct interface index position.
@@ -72,18 +129,18 @@ def update_interfaces(total_interfaces, interface_list):
     return updated_interface_list
 
 
-def update_guests(guests):
+def update_guest_interfaces(guest_data):
     """
     Entrypoint to updating guest data parameters.
-    :param guests: List of host dicts.
+    :param guest_data: List of host dicts.
     :return: New list of host dicts.
     """
     updated_guest_list = []
-    for guest in guests:
+    for guest in guest_data:
         if not guest.get('interfaces'):
             updated_guest_list.append(guest)
         else:
-            guest['interfaces'] = update_interfaces(
+            guest['interfaces'] = add_blackhole_interfaces(
                 guest['provider_config']['nic_adapter_count'], guest['interfaces']
             )
             updated_guest_list.append(guest)
